chore: Api catalog functional tests infrastructure (#1803)
* First take on test


* Initial fail with first test


* Decouple tests from each other and enable Functional tests


* test name refactor


* Align config files for trustStore params
