+++ api-catalog-package/src/main/resources/bin/start.sh
@@ -98,7 +98,7 @@ _BPX_JOBNAME=${ZOWE_PREFIX}${CATALOG_CODE} java \
     -Dapiml.service.ipAddress=${ZOWE_IP_ADDRESS:-127.0.0.1} \
     -Dapiml.service.preferIpAddress=${APIML_PREFER_IP_ADDRESS:-false} \
     -Dapiml.service.gatewayHostname=${GATEWAY_HOST} \
-    -Dapiml.service.eurekaUserId=eureka \
+    -Dapiml.service.eurekaUserName=eureka \
     -Dapiml.service.eurekaPassword=password \
     -Dapiml.logs.location=${WORKSPACE_DIR}/api-mediation/logs \
     -Dapiml.discovery.staticApiDefinitionsDirectories=${APIML_STATIC_DEF} \
