+++ api-catalog-services/src/main/java/org/zowe/apiml/apicatalog/ApiCatalogApplication.java
@@ -9,9 +9,6 @@
  */
 package org.zowe.apiml.apicatalog;
 
-import org.zowe.apiml.product.logging.annotations.EnableApimlLogger;
-import org.zowe.apiml.product.monitoring.LatencyUtilsConfigInitializer;
-import org.zowe.apiml.product.version.BuildInfo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
@@ -20,10 +17,13 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import org.zowe.apiml.product.logging.annotations.EnableApimlLogger;
+import org.zowe.apiml.product.monitoring.LatencyUtilsConfigInitializer;
+import org.zowe.apiml.product.version.BuildInfo;
 
 @SpringBootApplication(exclude = HystrixAutoConfiguration.class)
 @EnableEurekaClient
-@ComponentScan({
+@ComponentScan(value = {
     "org.zowe.apiml.apicatalog",
     "org.zowe.apiml.product.security",
     "org.zowe.apiml.product.web",
