+++ api-catalog-services/src/test/java/org/zowe/apiml/apicatalog/functional/PrototypeFunctionalTest.java
@@ -1,0 +1,22 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+package org.zowe.apiml.apicatalog.functional;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+class PrototypeFunctionalTest extends ApiCatalogFunctionalTest {
+
+    @Test
+    void testContextStarts() {
+        assertTrue(true);
+    }
+}
