+++ api-catalog-services/src/test/java/org/zowe/apiml/apicatalog/instance/InstanceServicesContextConfiguration.java
@@ -1,0 +1,21 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+
+package org.zowe.apiml.apicatalog.instance;
+
+import org.springframework.context.annotation.Bean;
+import org.zowe.apiml.apicatalog.discovery.DiscoveryConfigProperties;
+
+public class InstanceServicesContextConfiguration {
+    @Bean
+    public DiscoveryConfigProperties discoveryConfigProperties() {
+        return new DiscoveryConfigProperties();
+    }
+}
