+++ api-catalog-services/src/test/java/org/zowe/apiml/apicatalog/staticapi/StaticAPIRefreshControllerTest.java
@@ -14,18 +14,14 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.security.config.annotation.web.WebSecurityConfigurer;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.client.RestClientException;
 import org.zowe.apiml.apicatalog.services.status.model.ServiceNotFoundException;
-import org.zowe.apiml.message.core.MessageService;
-import org.zowe.apiml.message.yaml.YamlMessageService;
 
 import static org.hamcrest.Matchers.hasSize;
 import static org.mockito.Mockito.when;
@@ -38,6 +34,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
     excludeFilters = { @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, value = WebSecurityConfigurer.class) },
     excludeAutoConfiguration = { SecurityAutoConfiguration.class}
 )
+@ContextConfiguration(classes = StaticApiContextConfiguration.class)
 class StaticAPIRefreshControllerTest {
 
     private static final String API_REFRESH_ENDPOINT = "/static-api/refresh";
@@ -88,25 +85,4 @@ class StaticAPIRefreshControllerTest {
             .andExpect(status().isOk());
     }
 
-    @Configuration
-    static class ContextConfiguration {
-
-        @MockBean
-        private StaticAPIService staticAPIService;
-
-        @Bean
-        public MessageService messageService() {
-            return new YamlMessageService("/apicatalog-log-messages.yml");
-        }
-
-        @Bean
-        public StaticAPIRefreshControllerExceptionHandler staticAPIRefreshControllerExceptionHandler() {
-            return new StaticAPIRefreshControllerExceptionHandler(messageService());
-        }
-
-        @Bean
-        public StaticAPIRefreshController apiCatalogController() {
-            return new StaticAPIRefreshController(staticAPIService);
-        }
-    }
 }
