+++ config/local-multi/api-catalog-service-1.yml
@@ -24,7 +24,7 @@ apiml:
         discoveryServiceUrls: https://localhost2:10021/eureka/,https://localhost:10011/eureka/
         dsIpAddress: 0.0.0.0
         eurekaPassword: password
-        eurekaUserId: eureka
+        eurekaUserName: eureka
         gatewayHostname: https://localhost:10010
         preferIpAddress: false
         truststore: keystore/localhost/localhost-multi.truststore.p12
