+++ config/zowe-api-dev/alternative_start.sh
@@ -158,7 +158,7 @@ _BPX_JOBNAME=${ZOWE_PREFIX}${CATALOG_CODE} java \
     -Dapiml.service.ipAddress=${ZOWE_IP_ADDRESS} \
     -Dapiml.service.preferIpAddress=${APIML_PREFER_IP_ADDRESS} \
     -Dapiml.service.gatewayHostname=${ZOWE_EXPLORER_HOST} \
-    -Dapiml.service.eurekaUserId=eureka \
+    -Dapiml.service.eurekaUserName=eureka \
     -Dapiml.service.eurekaPassword=password \
     -Dapiml.logs.location=${WORKSPACE_DIR}/api-mediation/logs \
     -Dapiml.security.ssl.verifySslCertificatesOfServices=${VERIFY_CERTIFICATES:-false} \
